move code to private module...
... in skrifa. Completely replaces the old TrueType scaler.
