+++ read-fonts/src/tables.rs
@@ -26,7 +26,6 @@ pub mod os2;
 pub mod post;
 pub mod postscript;
 pub mod stat;
-pub mod truetype;
 pub mod variations;
 pub mod vhea;
 pub mod vmtx;
