+++ skrifa/src/scale/cff/mod.rs
@@ -1,8 +1,5 @@
 //! Support for scaling CFF outlines.
 
-// Temporary until new scaler API is done.
-#![allow(dead_code)]
-
 mod hint;
 mod scaler;
