+++ skrifa/src/scale/error.rs
@@ -12,10 +12,11 @@ pub enum Error {
     NoSources,
     /// The requested glyph was not present in the font.
     GlyphNotFound(GlyphId),
+    /// Exceeded memory limits when loading a glyph.
+    InsufficientMemory,
     /// Exceeded a recursion limit when loading a glyph.
     RecursionLimitExceeded(GlyphId),
     /// Error occured during hinting.
-    #[cfg(feature = "hinting")]
     HintingFailed(GlyphId),
     /// An anchor point had invalid indices.
     InvalidAnchorPoint(GlyphId, u16),
@@ -50,12 +51,12 @@ impl fmt::Display for Error {
         match self {
             Self::NoSources => write!(f, "No glyph sources are available for the given font"),
             Self::GlyphNotFound(gid) => write!(f, "Glyph {gid} was not found in the given font"),
+            Self::InsufficientMemory => write!(f, "exceeded memory limits"),
             Self::RecursionLimitExceeded(gid) => write!(
                 f,
                 "Recursion limit ({}) exceeded when loading composite component {gid}",
                 super::GLYF_COMPOSITE_RECURSION_LIMIT,
             ),
-            #[cfg(feature = "hinting")]
             Self::HintingFailed(gid) => write!(f, "Bad hinting bytecode for glyph {gid}"),
             Self::InvalidAnchorPoint(gid, index) => write!(
                 f,
